--- client/src/pages/Footer.js
@@ -1,39 +1,0 @@
-import React from 'react';
-
-function Footer() {
-    return (
-<   div className="main-footer">
-        <div className="container">
-            <div className="row">
-             <div className="col-md-6">
-                <h4>Our Mission</h4>
-                    <p> The Goal of "While You're There is to (describe goal and mission, mention reduced store trips and covid safety application
-
-                    </p>
-            </div>
-         <div className="col-md-6">
-                <h4>We Love Feedbeack!</h4>
-                <ul className="list-unstyled">
-                    <li>email </li>
-                    <li>email</li>
-                    <li>email</li>
-                    <li>email</li>
-                    
-                </ul>
-            </div>
-
-<div className="footer-bottom">
-    <p className="text-xs-center">
-     &copy;{new Date().getFullYear()} While You're There App 
-     </p>
-</div>
-  </div>
-  </div>
-        
-  </div>
-
-
-    );
-}
-
-export default Footer;
