+++ client/src/pages/Home.jsx
@@ -3,30 +3,35 @@ import React, { useEffect, useState } from "react";
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Button from 'react-bootstrap/Button';
-
+import Container from 'react-bootstrap/Container';
 
 
 function Home() {
     return (
-        <div>
+        <Container>
             <Row className="teams justify-content-center">
-                <Col sm="auto" className="mb-2">
+                <Col sm="auto" className="t mb-3">
                     <Button variant="primary" type="submit" id="createTeam">Create Group</Button>
                 </Col>
-                <Col sm="auto" className="mb-2">
+             </Row>
+             <Row className="teams justify-content-center">
+                <Col sm="auto" className="mb-3">
                     <Button variant="primary" type="submit" id="viewTeams">View Groups</Button>
                 </Col>
             </Row>
             <Row className="teams justify-content-center">
-                <Col sm="auto" className="mb-2">
+                <Col sm="auto" className="mb-3">
                     <Button variant="primary" type="submit" id="updateUserProfile">Update User Profile</Button>
                 </Col>
-                <Col sm="auto" className="mb-2">
+             </Row>
+             <Row className="teams justify-content-center">
+                <Col sm="auto" className="mb-3">
                     <Button variant="primary" type="submit" id="viewGroupStores">View Group Stores</Button>
                 </Col>
+                
             </Row>
 
-        </div>
+        </Container>
 
     )
 }
